split multiple classes
